+++ src/main/java/insane96mcp/iguanatweaksreborn/module/client/Light.java
@@ -16,6 +16,9 @@ public class Light extends Feature {
     @Config
     @Label(name = "No Night Vision Flashing", description = "If true night vision will no longer flash 10 seconds before expiring, instead will slowly fade out 2 seconds before expiring.")
     public static Boolean noNightVisionFlashing = true;
+    @Config
+    @Label(name = "Force darkness", description = "If enabled, the brightness will be set to 0 regardless of the current brightness in video settings.")
+    public static Boolean forceDarkness = true;
 
     public Light(Module module, boolean enabledByDefault, boolean canBeDisabled) {
         super(module, enabledByDefault, canBeDisabled);